make Signals a final class
